# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Question Queue member API
  description: |
    A Flask + MongoDB API for the member service.
    
    ## Authentication
    All endpoints (except `/dev-login` and `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a token from the `/dev-login` endpoint.
  version: 0.1.0
  contact:
    name: Stage0 Demo Team
    email: info@stage0-demo-team.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8292
    description: Local development server
...
paths:
...
  /api/Curriculum:
    get:
      tags:
        - Curriculum
      summary: Get all Curriculum documents
      description: Retrieve infinite scroll batch of sorted, filtered Curriculum documents
      operationId: getCurriculums
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Curriculum
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Curriculum
      summary: Create a new Curriculum document
      description: Create a new Curriculum document (requires staff or admin role)
      operationId: createCurriculum
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurriculumInput'
      responses:
        '201':
          description: Curriculum created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Curriculum/{CurriculumId}:
    get:
      tags:
        - Curriculum
      summary: Get Curriculum by CurriculumId
      description: Retrieve a specific Curriculum document by its CurriculumId
      operationId: getCurriculum
      security:
        - bearerAuth: []
      parameters:
        - name: CurriculumId
          in: path
          required: true
          description: The Curriculum ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Curriculum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curriculum'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Curriculum
      summary: Update Curriculum document
      description: Update a Curriculum document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateCurriculum
      security:
        - bearerAuth: []
      parameters:
        - name: CurriculumId
          in: path
          required: true
          description: The Curriculum ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurriculumUpdate'
      responses:
        '200':
          description: Curriculum updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curriculum'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/Review:
    get:
      tags:
        - Review
      summary: Get all Review documents
      description: Retrieve infinite scroll batch of sorted, filtered Review documents
      operationId: getReviews
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Review
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Review
      summary: Create a new Review document
      description: Create a new Review document (requires staff or admin role)
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Review/{ReviewId}:
    get:
      tags:
        - Review
      summary: Get Review by ReviewId
      description: Retrieve a specific Review document by its ReviewId
      operationId: getReview
      security:
        - bearerAuth: []
      parameters:
        - name: ReviewId
          in: path
          required: true
          description: The Review ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Review
      summary: Update Review document
      description: Update a Review document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateReview
      security:
        - bearerAuth: []
      parameters:
        - name: ReviewId
          in: path
          required: true
          description: The Review ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdate'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/dev-login` endpoint.
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
    Curriculum:
      type: object
      description: Curriculum domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Curriculum name
          pattern: '^[^\s]{1,40}$'
          example: my-Curriculum
        description:
          type: string
          description: Curriculum description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Curriculum document
        status:
          type: string
          description: Curriculum status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    CurriculumInput:
      type: object
      description: Input for creating a Curriculum document
      properties:
        name:
          type: string
          description: Curriculum name
          pattern: '^[^\s]{1,40}$'
          example: my-Curriculum
        description:
          type: string
          description: Curriculum description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Curriculum document
        status:
          type: string
          description: Curriculum status
          enum:
            - active
            - archived
          example: active
    CurriculumUpdate:
      type: object
      description: Fields to update in a Curriculum document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Curriculum name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Curriculum description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Curriculum status
          enum:
            - active
            - archived
...
    Review:
      type: object
      description: Review domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Review name
          pattern: '^[^\s]{1,40}$'
          example: my-Review
        description:
          type: string
          description: Review description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Review document
        status:
          type: string
          description: Review status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    ReviewInput:
      type: object
      description: Input for creating a Review document
      properties:
        name:
          type: string
          description: Review name
          pattern: '^[^\s]{1,40}$'
          example: my-Review
        description:
          type: string
          description: Review description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Review document
        status:
          type: string
          description: Review status
          enum:
            - active
            - archived
          example: active
    ReviewUpdate:
      type: object
      description: Fields to update in a Review document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Review name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Review description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Review status
          enum:
            - active
            - archived
...